[chore] cleanup type definitions
